Quote and base controller done

+++ database/migrations/2022_12_26_103201_create_quotations_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->string('status');
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
-            $table->integer('consignee_id');
+            $table->unsignedBigInteger('consignee_id');
             $table->foreign('consignee_id')->references('id')->on('consignments');
             $table->string('quotation_number');
             $table->float('sub_total');
